question-detector/app/requirements.txt
pillow
tqdm
jupyter
ipykernel
fastapi
uvicorn
pillow
ultralytics
python-multipart